requirements.txt
opencv-python==4.7.0.72
pillow==12.0.1
pytesseract==0.3.13
numpy==1.26.0
pyautogui==0.9.53
argostranslate==1.9.1
